Make useTrackPlayer hook to play track in chat
