+++ packages/common/src/hooks/chats/index.ts
@@ -1,2 +1,3 @@
 export * from './useCanSendMessage'
 export * from './useSetInboxPermissions'
+export * from './useTrackPlayer'
