+++ packages/mobile/src/screens/chat-screen/ChatMessageTrack.tsx
@@ -1,13 +1,11 @@
 import { TrackTile } from 'app/components/lineup-tile'
-import { Kind, Name, PlaybackSource, QueueSource, accountSelectors, getPathFromTrackUrl, makeUid, playerSelectors, queueActions, queueSelectors, useGetTrackByPermalink } from "@audius/common"
-import { useDispatch, useSelector } from "react-redux"
+import { Kind, PlaybackSource, QueueSource, accountSelectors, getPathFromTrackUrl, makeUid, useGetTrackByPermalink, useTrackPlayer } from "@audius/common"
+import { useSelector } from "react-redux"
 import { useCallback, useMemo } from "react"
-import { make } from 'common/store/analytics/actions'
 
 const { getUserId } = accountSelectors
-const { makeGetCurrent } = queueSelectors
-const { getPlaying } = playerSelectors
-const { clear, add, play, pause } = queueActions
+import { make, track as trackEvent } from 'app/services/analytics'
+import { EventNames } from 'app/types/analytics'
 
 type ChatMessageTrackProps = {
   link: string
@@ -15,10 +13,7 @@ type ChatMessageTrackProps = {
 }
 
 export const ChatMessageTrack = ({ link, isAuthor }: ChatMessageTrackProps) => {
-  const dispatch = useDispatch()
   const currentUserId = useSelector(getUserId)
-  const currentQueueItem = useSelector(makeGetCurrent())
-  const playing = useSelector(getPlaying)
 
   const permalink = getPathFromTrackUrl(link)
   const { data: track, status } = useGetTrackByPermalink(
@@ -41,56 +36,39 @@ export const ChatMessageTrack = ({ link, isAuthor }: ChatMessageTrackProps) => {
   } : null
 
   const uid = useMemo(() => {
-    return track ? makeUid(Kind.TRACKS, track.track_id) : ''
+    return track ? makeUid(Kind.TRACKS, track.track_id) : null
   }, [track])
-  const isTrackPlaying =
-    playing &&
-    !!track &&
-    !!currentQueueItem.track &&
-    currentQueueItem.uid === uid
 
-  const recordAnalytics = useCallback(
-    ({ name, source }: { name: Name; source: PlaybackSource }) => {
-      if (!track) return
-      dispatch(
-        make(name, {
-          id: `${track.track_id}`,
-          source
-        })
-      )
-    },
-    [dispatch, track]
-  )
-
-  const onTogglePlay = useCallback(() => {
-    if (!track) return
-    if (isTrackPlaying) {
-      dispatch(pause({}))
-      recordAnalytics({
-        name: Name.PLAYBACK_PAUSE,
+  const recordPlay = useCallback(() => {
+    trackEvent(
+      make({
+        eventName: EventNames.PLAYBACK_PLAY,
         source: PlaybackSource.CHAT_TRACK
       })
-    } else {
-      dispatch(clear({}))
-      dispatch(
-        add({
-          entries: [
-            { id: track.track_id, uid, source: QueueSource.CHAT_TRACKS }
-          ]
-        })
-      )
-      dispatch(play({ uid }))
-      recordAnalytics({
-        name: Name.PLAYBACK_PLAY,
+    )
+  }, [])
+
+  const recordPause = useCallback(() => {
+    trackEvent(
+      make({
+        eventName: EventNames.PLAYBACK_PAUSE,
         source: PlaybackSource.CHAT_TRACK
       })
-    }
-  }, [dispatch, recordAnalytics, track, isTrackPlaying, currentQueueItem, uid])
+    )
+  }, [])
 
-  return item && user ? (
+  const { togglePlay } = useTrackPlayer({
+    id: track?.track_id ?? null,
+    uid,
+    queueSource: QueueSource.CHAT_TRACKS,
+    recordPlay,
+    recordPause
+  })
+
+  return item && user && uid ? (
     <TrackTile
       index={0}
-      togglePlay={onTogglePlay}
+      togglePlay={togglePlay}
       uid={uid}
       isTrending={false}
       showArtistPick={false}
